+++ src/main/java/com/blackbaud/events/core/domain/TicketRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.util.List;
 
-public interface TicketRepository extends PagingAndSortingRepository<TicketEntity, Long> {
+public interface TicketRepository extends PagingAndSortingRepository<TicketEntity, Integer> {
 
     List<TicketEntity> findByEventId(Integer eventId);
